# Issue #11727: Include CDG Accelerator Plugin to boost pitest performance