--- a/.ci/jsoref-spellchecker/whitelist.words
+++ b/.ci/jsoref-spellchecker/whitelist.words
@@ -528,6 +528,7 @@ Grenner
 grep
 grepped
 groovyism
+groupcdg
 groupon
 grp
 Gsp