@Vyom-Yadav 2022-07-09
> pitest-annotation is 59s - https://github.com/checkstyle/checkstyle/runs/7264372519?check_suite_focus=true but in your report it is 1.44 . do you provide numbers from your local execution ?
> 
> it make sense to do comparison of Github Action execution, as it will run it all the time.

I provided the total execution time, not just the time of generating the report. Basically, it was easier to copy from https://github.com/checkstyle/checkstyle/actions/runs/2641420843/usage rather than manually typing every step execution time.