@Vyom-Yadav 2022-07-09
> it would be awesome to make one more bar at https://docs.arcmutate.com/.
> We can try just for fun make single profile and run it on local.
> Or we can measure how much build takes time before pitest execution and reduce this time from each launch and sum up "pure" pitest execution.
> as we have this number, we can share it back to arcmutate, to let them be proud of their project.

Yes, of course, I will make 2 PRs, both with a single profile, that runs the complete pitest (ant, coding, ....). One with the accelerator plugin and the other without it, then we can compare the time and send the report to arcmutate. They are doing amazing work for the open-source community.