@romani 2022-07-09
>  I will make 2 PRs

be careful with this .... last time we did pitest as single run .... it took 12 hours to pass. I am not sure if same bad will be now, but be careful, run on local first.