@romani 2022-07-09
https://github.com/checkstyle/checkstyle/runs/7264372538?check_suite_focus=true 1m run.

if we spend 20 seconds on start/finish . and from 40 secnds of mvn execution, about 10 goes to compilation. So we spend 50% on non-pistest ...... it is time to rethink split of pitest execution and make them to run for ~5 minutes to spedn more on pitest rather then start/finish