@nick-mancuso 2022-07-09
> it would be awesome to make one more bar at https://docs.arcmutate.com/.

Agreed, but let's just create an issue on this, and come back to it after we use this plugin for awhile to let us gain some experience with the plugin and allow our new pitest execution/ report analysis to gain some maturity.

> if we spend 20 seconds on start/finish . and from 40 secnds of mvn execution, about 10 goes to compilation. So we spend 50% on non-pistest ...... it is time to rethink split of pitest execution and make them to run for ~5 minutes to spedn more on pitest rather then start/finish

Agreed, we can probably combine a few quicker profiles, @Vyom-Yadav let's create an issue on this and think of some logical grouping.